wanghong/requirements.txt
# 项目依赖

# 网络请求
aiohttp>=3.9.0
aiofiles>=23.0.0
requests>=2.31.0

# X 平台爬虫（使用 twscrape 或其他库）
# twscrape>=0.12.0

# AI 相关
openai>=1.0.0
litellm>=1.0.0

# 数据处理
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 工具
pillow>=10.0.0  # 图片处理
tqdm>=4.66.0    # 进度条
rich>=13.0.0    # 美化输出

# 可选：视频处理
# ffmpeg-python>=0.2.0
